added flag image type and acquisition phase

+++ matlab/toolbox/+mrd/+binary/ImageHeaderSerializer.m
@@ -21,7 +21,7 @@ classdef ImageHeaderSerializer < yardl.binary.RecordSerializer
       field_serializers{16} = yardl.binary.Uint64Serializer;
       field_serializers{17} = yardl.binary.Uint64Serializer;
       field_serializers{18} = yardl.binary.EnumSerializer('mrd.ImageType', @mrd.ImageType, yardl.binary.Int32Serializer);
-      field_serializers{19} = yardl.binary.OptionalSerializer(yardl.binary.EnumSerializer('mrd.ImageQuantitativeType', @mrd.ImageQuantitativeType, yardl.binary.Int32Serializer));
+      field_serializers{19} = yardl.binary.OptionalSerializer(yardl.binary.EnumSerializer('mrd.ImageQuantitativeType', @mrd.ImageQuantitativeType, yardl.binary.Uint64Serializer));
       field_serializers{20} = yardl.binary.OptionalSerializer(yardl.binary.Uint32Serializer);
       field_serializers{21} = yardl.binary.OptionalSerializer(yardl.binary.Uint32Serializer);
       field_serializers{22} = yardl.binary.VectorSerializer(yardl.binary.Int32Serializer);
